解决maven install出现的<T>问题

+++ quickbundle-rmwebdemo/src/main/java/org/quickbundle/modules/message/dao/RmMessageDao.java
@@ -123,7 +123,7 @@ public class RmMessageDao extends RmSqlSessionDaoSupport implements IRmMessageCo
      * @return 总记录数
      */
     public int getCount(String queryCondition) {
-    	return getSqlSession().selectOne(namespace("getCount"), queryCondition);
+    	return (Integer)getSqlSession().selectOne(namespace("getCount"), queryCondition);
     }
     
     /**
